requirements.txt
pip==22.3.1
setuptools==65.6.3
wheel==0.38.4

python-dotenv~=0.21.0